# Add new example with faster field propagation