This is based on Example 13, but restricts the step length of charged
particles in case of magnetic field for better performance. The idea
is to make shorter steps where propagation finishes after a low-ish
number of chord steps to reduce divergence. The step limit has to
happen *before* MSC to avoid approximations when just aborting field
propagation (as was already possible with the max_iterations argument).

Performance numbers on RTX 2070 SUPER, with `-particles 1000`:
```
  Bz = 1T
    example13: 24.74s, example18: 23.10s, -6.7%
  Bz = 3.8T
    example13: 96.04s, example18: 79.46s, -17.3%
```